category controller merge changes
