+++ src/main/java/org/upgrad/services/CategoryService.java
@@ -1,13 +1,20 @@
 package org.upgrad.services;
 
-import org.springframework.stereotype.Service;
 import org.upgrad.models.Category;
 
-@Service
+/*
+ * This CategoryService interface gives the list of all the service that exist in the category service implementation class.
+ * Controller class will be calling the service methods by this interface.
+ */
 public interface CategoryService {
 
+    // Get all category details.
     Iterable<Category> getAllCategories();
 
-    Category getCategoryByName(String name);
+    //Get category details by name.
+    Category getCategory(String categoryName);
+
+    //Get category count by name.
+    Integer getCategoryCount(String categoryName);
 
 }
